utmp: move `failtmp()`
The functionality from this function is related to btmp.
