+++ libmisc/failure.h
@@ -13,7 +13,6 @@
 
 #include "defines.h"
 #include "faillog.h"
-#include <utmp.h>
 
 /*
  * failure - make failure entry
@@ -40,14 +39,6 @@ extern int failcheck (uid_t uid, struct faillog *fl, bool failed);
  *	message which is displayed at login time.
  */
 extern void failprint (const struct faillog *);
-
-/*
- * failtmp - update the cumulative failure log
- *
- *	failtmp updates the (struct utmp) formatted failure log which
- *	maintains a record of all login failures.
- */
-extern void failtmp (const char *username, const struct utmp *);
 
 #endif
